requirements.txt
colorama~=0.4.4
pillow~=9.1.1
pyautogui~=0.9.53
selenium~=4.1.5
bs4~=0.0.1

requests~=2.27.1
beautifulsoup4~=4.11.1